added a Makefile

+++ problem_15.cpp
@@ -21,13 +21,13 @@ int main () {
     vector<vector<long long>> matrix;
     matrix.resize(21, vector<long long>(21));
 
-    for (int i = 0; i < matrix.size(); ++i) {
+    for (size_t i = 0; i < matrix.size(); ++i) {
         matrix[0][i] = 1;
         matrix[i][0] = 1;
     }
 
-    for (int row = 1; row < matrix.size(); ++row) {
-        for (int col = 1; col < matrix.size(); ++col) {
+    for (size_t row = 1; row < matrix.size(); ++row) {
+        for (size_t col = 1; col < matrix.size(); ++col) {
             matrix[row][col] = matrix[row - 1][col] + matrix[row][col - 1];
         }
     }
